Made two separate files for testing live and csvs that now use assert_test and log hyperparameters. Also updated README.MD to reflect these changes.
